backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ML and NLP
spacy==3.7.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2

# PDF/Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# Google Cloud Platform
google-cloud-storage==2.10.0
google-cloud-sql-connector==1.4.3
google-cloud-pubsub==2.18.4
google-cloud-monitoring==2.16.0
google-cloud-logging==3.8.0
google-cloud-aiplatform==1.38.1

# Redis for caching
redis==5.0.1
hiredis==2.2.3

# HTTP and utilities
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Monitoring and observability
prometheus-client==0.19.0
structlog==23.2.0

# Security
cryptography==41.0.8

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production
gunicorn==21.2.0